# [Relay][TOPI] Remove input padding for arm_cpu conv2d int8 native schedule in Legalize pass