The Legalize pass was unnecessarily padding the input channels for conv2d int8 native implementations. Since the conv2d schedule itself can add padding more efficiently, I skipped padding during the pass and further optimised the schedule to deal with it.  
I also added a test to check whether or not the Legalize pass pads the conv2d input data.  

The benchmark results of a single int8 conv2d operation where input padding along the K axis is necessary, with input = (1 x 224 x 224 x 3), filter = (16 x 3 x 3 x 3), output = (1 x 112 x 112 x 16), tested for the following target `"llvm --device=arm_cpu --mtriple=aarch64-linux-gnu -mattr=+v8.2a,+dotprod"` , before and after the changes, is given below:

|  Before (ms) | After (ms) | Speedup (%) |
|-------------|------------|-------------|
| 0.5186      | 0.0663     | 87.22       |