@Anndrey24 2023-09-05
Regarding the tests, it seems that all of the padding cases are already visited at least once in: https://github.com/apache/tvm/blob/476d9d1d41980ad35cd6a4d044c17fea2210f352/tests/python/topi/python/test_topi_conv2d_int8.py#L110